Partially move type unify into a separate module
The move is partial because it failed to move unify, but only moved
its related functions. More functions related to fresh type vars should
be moved as well. Occurs check should be an internal part of
aeso_type_unify and should not be exported after unify is moved to the
mentioned module.