Type checker refactoring (#526)
* Move ets-related functions into a separate module

* Move typing related modules into typing directory

* Move pretty printing partially to aeso_type_pretty

The move is partial because pp_when and pp_why_record depend on other
functions and records that should not be moved into aeso_type_pretty.

This commit also duplicates the function typesig_to_fun_t, so it now
exists in both aeso_ast_infer_types.erl and aeso_type_pretty.erl.

* Move pp_why_record into a separate module

This change is temporary. The function pp_why_record must end up in the
module aeso_type_pretty. This change causes the duplication of the type
definition why_record() as well as the function pos/1.

* Partially move type unify into a separate module

The move is partial because it failed to move unify, but only moved
its related functions. More functions related to fresh type vars should
be moved as well. Occurs check should be an internal part of
aeso_type_unify and should not be exported after unify is moved to the
mentioned module.

* Move pp_when into a separate module

This move has resulted in the duplication of the record
named_argument_constraint. It also includes duplication of the helper
functions pos/1 and pos/2.

* Move mk_error away from aeso_ast_infer_types

The new module contains duplicated funs and type defs

* Move warnings creation into a separate module

* Move duplicated types and records into header file

* Remove the duplication of why_record() type

* Merge all pretty printing modules

* Move env record with all related type/records into header file

* Move global_env and its related functions into a separate module

* Move get_current_scope into aeso_type_env

* Move type error management functions into aeso_type_errors

This has resulted in the duplication of the functions qid and qcon.

* Move warnings creators into a separate module

Helper functions were duplicated: name, qname, and get_option.

* Move type env related function into aeso_type_env

The function bind_consts is dependent on inference and could not be moved

* Move types unfolding into a separate module

* Move unification to aeso_type_unify

* Move init_env into aeso_type_env

* Merge warnings modules

* Rename aeso_infer_ets to aeso_type_ets

* Rename aeso_ast_infer_types to aeso_type_infer

* Move more helpers to the helper funs module

* Move type_error/1 to helper functions

* Make calls to aeso_type_env outside of warnings module

* Move when_warning into aeso_type_env

This might not be the right place to have that function. The function
was moved in order to break circular dependency between modules.

For now, it seems that this is the right place for when_warning. Later
the function could be changed, or it could be moved somewhere else.

* Move more unfolding function into aeso_type_unfold

* Move more helper function to aeso_type_helpers

* Move constraints solving into a new module

* Do not handle invalid warnings as type errors

This might not show a formatted warning and should be changed later but
will be left like that for the purpose of this commit.

* Split errors and warnings formatting

* Merge warnings and error formatting into one module

* Reduce code duplication

* Delete duplicated functioned from aeso_type_warnings

* Fix bad formatting

* Clean export list of aeso_type_ets

* Remove more duplicated functions

* Docs update

* Remove duplicated function

* More cleanup in env and infer

* Move when_warning to aeso_type_warnings

* Move unqualify to helper functions

* Move ets cleanup to aeso_type_ets

* Split desugaring into a separate module

* Docs update

* Rename aeso_desugar to aeso_type_desugar

* Formatting in aeso_type_desugar

* Remove unused includes in aeso_type_desugar

* Reduce code duplication and relocate more helpers